refactor: don't use feature config types as public types
